fix tests to match with new schema
